backend/requirements.txt
fastapi==0.129.0
uvicorn[standard]==0.40.0
SQLAlchemy==2.0.46
python-multipart==0.0.22
passlib==1.7.4
bcrypt==3.2.2
python-dotenv==1.2.1
opencv-python==4.13.0.92
opencv-contrib-python==4.13.0.92
numpy==2.2.6
mediapipe==0.10.9
pydantic[email]==2.12.5
PyJWT==2.11.0
email-validator==2.3.0